# src/docs/v1/openapi.yaml
openapi: 3.0.3
info:
  title: SyncBoard API
  version: "1.0.0"
  description: >
    Stateless REST API for auth, Workspaces, Boards, Lists, and Cards. Every
    route except /auth/register, /auth/login, /auth/refresh,
    /auth/forgot-password, and /auth/reset-password requires a bearer access
    token minted via /auth/register or /auth/login. Workspace/board/list/card
    routes additionally require the caller to be a member (Admin or Member)
    of the resource's workspace; a subset of workspace routes require the
    Admin role specifically.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
tags:
  - name: Auth
  - name: Workspaces
  - name: Boards
  - name: Lists
  - name: Cards

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Create an account and issue tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"

  /auth/login:
    post:
      tags: [Auth]
      summary: Exchange credentials for an access + refresh token pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthenticated"

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate a refresh token for a new access + refresh token pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPairEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthenticated"

  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current refresh token
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Unauthenticated"

  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset email (always accepted, no enumeration)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "202":
          description: Accepted
        "400":
          $ref: "#/components/responses/ValidationError"

  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Complete a password reset with a single-use token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/ValidationError"

  /workspaces:
    post:
      tags: [Workspaces]
      summary: Create a workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceEnvelope"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "400":
          $ref: "#/components/responses/ValidationError"
    get:
      tags: [Workspaces]
      summary: List the current user's workspaces
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceCollectionEnvelope"
        "401":
          $ref: "#/components/responses/Unauthenticated"

  /workspaces/{workspaceId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [Workspaces]
      summary: Get a single workspace
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Workspaces]
      summary: Update workspace name (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Workspaces]
      summary: Delete a workspace (Admin only; cascades boards/lists/cards)
      responses:
        "204":
          description: No Content
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /workspaces/{workspaceId}/members:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    post:
      tags: [Workspaces]
      summary: Add a member (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
                role:
                  type: string
                  enum: [Admin, Member]
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"

  /workspaces/{workspaceId}/members/{userId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - name: userId
        in: path
        required: true
        schema:
          type: string
    patch:
      tags: [Workspaces]
      summary: Update a member's role (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [Admin, Member]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Workspaces]
      summary: Remove a member (Admin only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /workspaces/{workspaceId}/boards:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    post:
      tags: [Boards]
      summary: Create a board in a workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [Boards]
      summary: List boards in a workspace
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardCollectionEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"

  /boards/{boardId}:
    parameters:
      - $ref: "#/components/parameters/BoardId"
    get:
      tags: [Boards]
      summary: Full board state (board + lists + cards)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      board:
                        $ref: "#/components/schemas/Board"
                      lists:
                        type: array
                        items:
                          $ref: "#/components/schemas/List"
                      cards:
                        type: array
                        items:
                          $ref: "#/components/schemas/Card"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Boards]
      summary: Update board title
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Boards]
      summary: Delete a board (cascades lists/cards)
      responses:
        "204":
          description: No Content
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /boards/{boardId}/lists:
    parameters:
      - $ref: "#/components/parameters/BoardId"
    post:
      tags: [Lists]
      summary: Create a list on a board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                order:
                  type: number
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [Lists]
      summary: List lists on a board
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListCollectionEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"

  /lists/{listId}:
    parameters:
      - $ref: "#/components/parameters/ListId"
    get:
      tags: [Lists]
      summary: Get a single list
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Lists]
      summary: Update title and/or order (reorder)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                order:
                  type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Lists]
      summary: Delete a list (cascades cards)
      responses:
        "204":
          description: No Content
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /lists/{listId}/cards:
    parameters:
      - $ref: "#/components/parameters/ListId"
    post:
      tags: [Cards]
      summary: Create a card in a list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                description:
                  type: string
                order:
                  type: number
                assignees:
                  type: array
                  items:
                    type: string
                labels:
                  type: array
                  items:
                    type: string
                checklist:
                  type: array
                  items:
                    type: object
                    properties:
                      text:
                        type: string
                      done:
                        type: boolean
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [Cards]
      summary: List cards in a list
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardCollectionEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"

  /cards/{cardId}:
    parameters:
      - $ref: "#/components/parameters/CardId"
    get:
      tags: [Cards]
      summary: Get a single card
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Cards]
      summary: >-
        Update content and/or move (listId + order). Moving into a list the
        caller cannot access returns 403.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                order:
                  type: number
                listId:
                  type: string
                assignees:
                  type: array
                  items:
                    type: string
                labels:
                  type: array
                  items:
                    type: string
                checklist:
                  type: array
                  items:
                    type: object
                    properties:
                      text:
                        type: string
                      done:
                        type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Cards]
      summary: Delete a card
      responses:
        "204":
          description: No Content
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
    Sort:
      name: sort
      in: query
      schema:
        type: string
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
    BoardId:
      name: boardId
      in: path
      required: true
      schema:
        type: string
    ListId:
      name: listId
      in: path
      required: true
      schema:
        type: string
    CardId:
      name: cardId
      in: path
      required: true
      schema:
        type: string

  responses:
    Unauthenticated:
      description: Missing or invalid access token
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: >
        Authenticated, but not a member (or not an Admin, where required) of
        the resource's workspace
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ValidationError:
      description: Request body failed validation
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Resource state conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string

    User:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time

    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string

    AuthEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            accessToken:
              type: string
            refreshToken:
              type: string

    TokenPairEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/TokenPair"

    WorkspaceMember:
      type: object
      properties:
        userId:
          type: string
        role:
          type: string
          enum: [Admin, Member]

    Workspace:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceMember"
        createdAt:
          type: string
          format: date-time

    Board:
      type: object
      properties:
        _id:
          type: string
        workspaceId:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time

    List:
      type: object
      properties:
        _id:
          type: string
        boardId:
          type: string
        title:
          type: string
        order:
          type: number

    ChecklistItem:
      type: object
      properties:
        text:
          type: string
        done:
          type: boolean

    Card:
      type: object
      properties:
        _id:
          type: string
        listId:
          type: string
        title:
          type: string
        description:
          type: string
        order:
          type: number
        assignees:
          type: array
          items:
            type: string
        labels:
          type: array
          items:
            type: string
        checklist:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"

    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    WorkspaceEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Workspace"

    WorkspaceCollectionEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Workspace"
        pagination:
          $ref: "#/components/schemas/Pagination"

    BoardEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Board"

    BoardCollectionEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Board"
        pagination:
          $ref: "#/components/schemas/Pagination"

    ListEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/List"

    ListCollectionEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/List"
        pagination:
          $ref: "#/components/schemas/Pagination"

    CardEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Card"

    CardCollectionEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Card"
        pagination:
          $ref: "#/components/schemas/Pagination"